Restructure settings
